Update itoa to handle 64bit and printf to handle hex

+++ libc/include/stdlib.h
@@ -2,6 +2,7 @@
 #define _STDLIB_H 1
 
 #include <sys/cdefs.h>
+#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -9,7 +10,7 @@ extern "C" {
 
 __attribute__((__noreturn__))
 void abort(void);
-char *itoa ( int value, char * str, int base );
+char *itoa ( int64_t value, char * str, int base );
 
 #ifdef __cplusplus
 }
